ENH Process the output of bedtools as a stream
The intermediate file (which was also written to the output) could be
quite large. This eliminates it and also may make the process faster.

On a test sample, wall clock time goes down from 14m21 to 9m (through
both parallelization [the output of bedtools is processed concurrently
with bedtools] as well as saving CPU time).