ENH Deprecate `--prodigal-output-faa` option
This is currently only partially implemented and it would require a
significant amount of work to do it properly (see
https://github.com/BigDataBiology/SemiBin/issues/185)

The reason for its introduction was that prodigal can be quite slow, so
it was becoming a bottleneck. Using this argument does replace calling
prodigal is parts of the pipeline (alas, not in every single instance,
hence this bug). However, in the meanwhile, we also introduced a much
faster ORF finding procedure (termed fast-naive) that just works as well
(for this intended case) and that is the default in SemiBin2. So, the
original reason for --prodigal-output-faa is moot now.

/close #185